fix(lmstudio): the download counters are fractional numbers
Found against a real LM Studio 0.4.20: a pull failed at 0 bytes with

  decode: json: cannot unmarshal number 5248034.973097618 into Go struct
  field downloadJob.bytes_per_second of type int64

LM Studio reports the download rate as a float, and the byte counts are the
same kind of number — the docs example happens to show integers. An int64
field fails the whole status document on it, so the poll never sees progress
and the job dies. The counters are float64 now, cast where they are used, and
a test pins the exact wire shape the real server sent.

+++ internal/backend/lmstudio/backend.go
@@ -211,7 +211,7 @@ func (b *Backend) Pull(ctx context.Context, req backend.PullRequest, progress fu
 		if progress == nil {
 			return
 		}
-		progress(backend.Progress{Status: status, BytesCompleted: j.DownloadedBytes, BytesTotal: j.TotalSizeBytes})
+		progress(backend.Progress{Status: status, BytesCompleted: int64(j.DownloadedBytes), BytesTotal: int64(j.TotalSizeBytes)})
 	}
 	if done, err := pullDone(job, ref); done || err != nil {
 		if err == nil {
@@ -275,7 +275,7 @@ func pullDone(job *downloadJob, ref string) (bool, error) {
 // describe words a progress event.
 func describe(job *downloadJob) string {
 	if job.BytesPerSecond > 0 {
-		return fmt.Sprintf("downloading (%s/s)", humanBytes(job.BytesPerSecond))
+		return fmt.Sprintf("downloading (%s/s)", humanBytes(int64(job.BytesPerSecond)))
 	}
 	return "downloading"
 }
